Add parameter of `deleteRoom`

+++ src/layers/services/room.service.js
@@ -1,4 +1,3 @@
-const { IoTSecureTunneling } = require('aws-sdk');
 const RoomRepository = require('../repositories/room.repository');
 
 class RoomService {
@@ -24,9 +23,10 @@ class RoomService {
         return await this.roomRepository.getCategoryRoom(category);
     };
 
-    deleteRoom = async (roomId) => {
+    deleteRoom = async (roomId, ownerUserId) => {
         const roomInfo = await this.roomRepository.getOneRoom(roomId);
         if (!roomInfo) return this.errResponse(400, '존재하지 않는 룸입니다.');
+        if (roomInfo.ownerUserId !== ownerUserId) return this.errResponse(401, '방장이 아닙니다.');
 
         const deleteRoom = await this.roomRepository.deleteRoom(roomId);
         if (!deleteRoom) return this.errResponse(401, '알 수 없는 에러');
